Added new tool: lcm_usarsim_proxy
The program lcm_usarsim_proxy makes a connection to the simulation environment USARSim, spawns a robot and publishes the sensor-messages to a number of lcm-channels. At the same time the program subscribes to the lcm-channel GAMEPAD, which allow you to drive the robot through the environment. USARSim is a simulation environment based on the Unreal Engine, which opens a number of tcp-ports to spawn different robot-models as GameBots. The program lcm_usarsim_proxy makes a connection to both the GamePort 3000 and the ImageServer port 5003. It polls both connections to check if data is available, and publishes this data on the lcm-channels POSE, LASER, CAMERA. This lcm-channel are broadcasted in the local subnetwork via the udpm-protocol on port 7667.

See for more information ./doc/index.html